+++ jest.config.js
@@ -8,6 +8,9 @@ const createJestConfig = nextJest({
 const customJestConfig = {
   moduleDirectories: ['node_modules', '<rootDir>/'],
   testEnvironment: 'jest-environment-jsdom',
+  transform: {
+    '\\.css\\.ts$': '@vanilla-extract/jest-transform',
+  },
 }
 
 module.exports = createJestConfig(customJestConfig)
