+++ package.json
@@ -25,6 +25,7 @@
     "@typescript-eslint/eslint-plugin": "^6.7.2",
     "@vanilla-extract/css": "^1.14.2",
     "@vanilla-extract/dynamic": "^2.1.0",
+    "@vanilla-extract/jest-transform": "^1.1.4",
     "@vanilla-extract/next-plugin": "^2.4.0",
     "@vanilla-extract/recipes": "^0.5.2",
     "autoprefixer": "10.4.16",
